move bin slicing test to slice_bins_test.py from bins_test.py and add undefined start/stop tests into existing tests
